pyproject.toml
[project]
name = "googleadk"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-cloud-aiplatform>=1.96.0",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "vertexai>=1.43.0",
]
